requirements.txt
streamlit
Pillow
pandas
numpy
matplotlib
streamlit_lottie
sklearn